Use fast-xml-parser over xq
In an effort to move backend jobs to github actions, we would like to make use
of this splendid action. Since our build machines did not have xq installed,
we could not run it out-of-the-box.

This should also make it easier to adapt this plugin for new environments
where we do not have xq installed.

I added a new dependency on fast-xml-parser to help with the xml parsing.